+++ griptape/tools/base_tool.py
@@ -18,6 +18,7 @@ from schema import Literal, Or, Schema
 from griptape.artifacts import BaseArtifact, ErrorArtifact, InfoArtifact, TextArtifact
 from griptape.common import observable
 from griptape.mixins.activity_mixin import ActivityMixin
+from griptape.mixins.runnable_mixin import RunnableMixin
 from griptape.mixins.serializable_mixin import SerializableMixin
 
 if TYPE_CHECKING:
@@ -27,7 +28,7 @@ if TYPE_CHECKING:
 
 
 @define
-class BaseTool(ActivityMixin, SerializableMixin, ABC):
+class BaseTool(ActivityMixin, SerializableMixin, RunnableMixin["BaseTool"], ABC):
     """Abstract class for all tools to inherit from for.
 
     Attributes:
@@ -122,11 +123,11 @@ class BaseTool(ActivityMixin, SerializableMixin, ABC):
 
         return schemas
 
-    def execute(self, activity: Callable, subtask: ActionsSubtask, action: ToolAction) -> BaseArtifact:
+    def run(self, activity: Callable, subtask: ActionsSubtask, action: ToolAction) -> BaseArtifact:
         try:
             output = self.before_run(activity, subtask, action)
 
-            output = self.run(activity, subtask, action, output)
+            output = self.try_run(activity, subtask, action, output)
 
             output = self.after_run(activity, subtask, action, output)
         except Exception as e:
@@ -135,10 +136,12 @@ class BaseTool(ActivityMixin, SerializableMixin, ABC):
         return output
 
     def before_run(self, activity: Callable, subtask: ActionsSubtask, action: ToolAction) -> Optional[dict]:
+        super().before_run()
+
         return action.input
 
     @observable(tags=["Tool.run()"])
-    def run(
+    def try_run(
         self,
         activity: Callable,
         subtask: ActionsSubtask,
@@ -163,6 +166,8 @@ class BaseTool(ActivityMixin, SerializableMixin, ABC):
         action: ToolAction,
         value: BaseArtifact,
     ) -> BaseArtifact:
+        super().after_run()
+
         if value:
             if self.output_memory:
                 output_memories = self.output_memory[getattr(activity, "name")] or []