+++ tests/unit/tools/test_base_tool.py
@@ -1,6 +1,7 @@
 import inspect
 import os
 import tempfile
+from unittest.mock import Mock
 
 import pytest
 from schema import Or, Schema, SchemaMissingKeyError
@@ -249,9 +250,9 @@ class TestBaseTool:
         assert MockTool().find_input_memory("foo") is None
         assert MockTool(input_memory=[defaults.text_task_memory("foo")]).find_input_memory("foo") is not None
 
-    def test_execute(self, tool):
+    def test_run(self, tool):
         action = ToolAction(input={}, name="", tag="")
-        assert tool.execute(tool.test_list_output, ActionsSubtask("foo"), action).to_text() == "foo\n\nbar"
+        assert tool.run(tool.test_list_output, ActionsSubtask("foo"), action).to_text() == "foo\n\nbar"
 
     def test_schema(self, tool):
         tool = MockTool()
@@ -309,7 +310,7 @@ class TestBaseTool:
         deserialized_tool = MockTool.from_dict(serialized_tool)
         assert isinstance(deserialized_tool, BaseTool)
 
-        assert deserialized_tool.execute(tool.test_list_output, ActionsSubtask("foo"), action).to_text() == "foo\n\nbar"
+        assert deserialized_tool.run(tool.test_list_output, ActionsSubtask("foo"), action).to_text() == "foo\n\nbar"
 
     def test_method_kwargs_var_injection(self, tool):
         tool = MockToolKwargs()
@@ -344,3 +345,13 @@ class TestBaseTool:
                 temp.seek(0)
 
                 assert InlineTool().are_requirements_met(temp.name) is True
+
+    def test_runnable_mixin(self, tool):
+        mock_on_before_run = Mock()
+        mock_after_run = Mock()
+        tool = MockTool(on_before_run=mock_on_before_run, on_after_run=mock_after_run)
+
+        tool.run(tool.test_list_output, ActionsSubtask("foo"), ToolAction(input={}, name="", tag="")).to_text()
+
+        mock_on_before_run.assert_called_once_with(tool)
+        mock_after_run.assert_called_once_with(tool)